Database/Database_connection.py
from pydantic import BaseModel,EmailStr , Field
from datetime import datetime
from typing import List, Annotated,Literal,Optional
import Database.Tables as Tables
from Database.database import engine, sessionlocal
from sqlalchemy.orm import Session

from fastapi import APIRouter, Depends, HTTPException, status
from passlib.context import CryptContext
from jose import JWTError, jwt

from Database.Tables import User, Builder, Project, ProjectImage,Agent

Tables.Base.metadata.create_all(bind=engine)

def get_db():
    db = sessionlocal()
    try:
        yield db
    finally:
        db.close()


db_dependency = Annotated[Session, Depends(get_db)]