requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0

pydantic==2.7.1
pydantic-settings==2.2.1

sqlalchemy==2.0.30

passlib[bcrypt]==1.7.4

python-jose[cryptography]==3.3.0
